pyproject.toml
[tool.poetry]
name = "iModels"
version = "0.1.0"
description = ""
authors = ["LinRenJie <xoxosos666@gmail.com>"]
readme = "README.md"
packages = [{ include = "app" }]

# 依赖管理
[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.1"
uvicorn = "^0.30.1"
openai-whisper = { git = "https://github.com/openai/whisper.git", rev = "v20231117" }
setuptools-rust = "^1.9.0"
numpy = "^1.26.4"
triton = { version = "2.3.1", optional = true }
openai = "^1.40.6"



# 其他依赖项
[tool.poetry.extras]
triton = ["triton"]

# 清华pypi源
[[tool.poetry.source]]
name = "mirrors"
url = "https://pypi.tuna.tsinghua.edu.cn/simple/"
priority = "primary"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"